Moves all .scss in a dedicated folder and deleted Carousel component
