+++ frontend/src/pages/ConfirmPage.jsx
@@ -1,10 +1,13 @@
-import "./ConfirmPage.scss";
+import "../styles/ConfirmPage.scss";
 import Navbar from "./Navbar";
 
 function ConfirmPage() {
   return (
     <>
-
+      <Navbar />
+      <p>Test</p>
     </>
   );
 }
+
+export default ConfirmPage;
