--- frontend/src/pages/Contact.scss
@@ -1,118 +1,0 @@
-$primary-color: #181818;
-$secondary-color: #151515;
-$text-color: #F9F9F9;
-$outline: #f9f9f9c5;
-
-body {
-    font-family: 'Inter', sans-serif;
-    background-color: $primary-color;
-    color: #F9F9F9;
-    margin: 0;
-}
-
-.contact_subtitle {
-    margin: 10px 0 50px 0;
-    font-weight: lighter;
-}
-
-.form {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    margin: 50px 0 50px 0;
-    background-color: $primary-color;
-    color: $text-color;
-}
-
-#sub_pos {
-    text-align: right;
-}
-
-.submit {
-    padding: 15px 40px 15px 40px;
-    background: #181818;
-    border: none;
-    outline: 1px solid $outline;
-    box-shadow: inset 0px 0px 20px rgba(255, 255, 255, 0.15);
-    border-radius: 18px;
-    font-size: large;
-    font-weight: bold;
-    color: $text-color;
-}
-
-.submit:hover {
-    box-shadow: inset 0px 0px 15px #f9f9f966;
-    outline: 1.5px solid $text-color;
-    padding: 12px 35px 12px 35px;
-    margin: 3px 5px;
-    cursor: pointer;
-}
-
-.input-group {
-    position: relative;
-    margin: 30px 0 30px 0;
-}
-.input-group_input {
-    padding: 10px;
-    border: none;
-    border-radius: 15px;
-    outline: 1.5px solid $outline;
-    background-color: transparent;
-    transition: outline-color 500ms;
-    width: 97.1%;
-    font-size: large;
-    color: $text-color;
-}
-.input-group_input:is(:focus, :valid) {
-    outline: 2px solid $text-color;
-}
-.input-group_label {
-    position: absolute;
-    pointer-events: none !important;
-    top: 0;
-    left: 0;
-    translate: 10px 10px;
-    color: $text-color;
-    transition: translate 500ms, scale 500ms;
-    background-color: $primary-color;
-    padding-inline: 10px;
-}
-.input-group_input:focus +
-.input-group_label,
-.input-group_input:valid +
-.input-group_label {
-    padding-inline: 10px;
-    translate: 5px -14px;
-    scale: 0.9;
-    background-color: $primary-color;
-}
-
-.input-group_input:focus +
-#name,
-.input-group_input:valid +
-#name {
-    margin-left: 3px;
-}
-
-#msg {
-    resize: none;
-}
-
-#msg::-webkit-scrollbar {
-    width: 10px;
-    margin: 10px;
-  
-    &-track {
-        background: $primary-color;
-        border-radius: 5px;
-    }
-  
-    &-thumb {
-        background: $outline;
-        border-radius: 5px;
-  
-    &:hover {
-        background: $text-color;
-      }
-    }
-  }
