+++ frontend/src/pages/MakeADrink.jsx
@@ -1,14 +1,11 @@
 import Navbar from "./Navbar";
-import MyCarousel from "../components/Carousel";
 
 function MakeADrink() {
   return (
     <div>
       <Navbar />
       <h2>Voici $VARIABLE cocktails à faire avec vos ingrédients</h2>
-      <div className="right-container">
-        <MyCarousel />
-      </div>
+      <div className="right-container" />
     </div>
   );
 }
