--- frontend/src/pages/MostPopular.scss
@@ -1,20 +1,0 @@
-.item {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 30vh;
-  height: 600px;
-  background-color: grey;
-  border-radius: 20px;
-  box-shadow:
-    rgba(0, 0, 0, 0.19) 0px 10px 20px,
-    rgba(0, 0, 0, 0.23) 0px 6px 6px;
-  overflow: hidden;
-  margin: 50px;
-
-  .elementContent {
-    display: flex;
-    flex-direction: column;
-    gap: 2rem;
-  }
-}
