+++ frontend/src/pages/Navbar.jsx
@@ -1,4 +1,4 @@
-import "./Navbar.scss";
+import "../styles/Navbar.scss";
 import logo from "../assets/logo.png";
 
 function Navbar() {
