+++ frontend/src/styles/MostPopular.scss
