+++ frontend/src/styles/Navbar.scss
